refactor: split transform manager modules
