+++ fold_node/src/fold_db_core/transform_manager/mod.rs
@@ -1,4 +1,6 @@
 pub mod manager;
+pub mod registry;
+pub mod executor;
 pub mod types;
 
 pub use manager::TransformManager;
